Split up rocksdb files
